requirements.txt
# Core Google ADK
google-genai

# Google AI
google-generativeai>=0.3.0

# LangChain for RAG
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.20

# Vector Database for RAG
chromadb>=0.4.22
faiss-cpu>=1.7.4

# Document Processing
PyPDF2>=3.0.1
pdfplumber>=0.10.3
python-docx>=1.1.0
markdown>=3.5.1
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Code Analysis
gitpython>=3.1.40
tree-sitter>=0.20.4
tree-sitter-python>=0.20.4
tree-sitter-javascript>=0.20.3
tree-sitter-java>=0.20.2
radon>=6.0.1
lizard>=1.17.10

# Web Scraping (for Confluence/Wiki)
requests>=2.31.0
atlassian-python-api>=3.41.0

# API Testing
requests>=2.31.0
httpx>=0.25.2

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pymongo>=4.6.1
redis>=5.0.1

# Testing Framework Generation
jinja2>=3.1.2

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
jsonschema>=4.20.0
tqdm>=4.66.1
rich>=13.7.0

# Web Framework (for API)
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.3

# CORS
fastapi-cors>=0.0.6

# File handling
aiofiles>=23.2.1

# Logging
loguru>=0.7.2
